+++ app/models/project.rb
@@ -66,7 +66,6 @@ class Project < ActiveRecord::Base
 
   has_many :contacts, -> { where deleted: false }
   has_many :documents, -> { where deleted: false }
-  has_many :links, -> { where deleted: false }
 
   has_many :domains, -> { where deleted: false }